+++ frontend/src/components/index.ts
@@ -1,0 +1,7 @@
+import AlertModal from './AlertModal/AlertModal';
+import InviteModal from './InviteModal/InviteModal';
+import Header from './layout/Header/Header';
+import RoomSettingModal from './RoomSettingModal/RoomSettingModal';
+import TopicContainer from './TopicContainer/TopicContainer';
+
+export { AlertModal, InviteModal, RoomSettingModal, TopicContainer, Header };
