+++ frontend/src/components/layout/Header/Header.stories.tsx
@@ -1,6 +1,7 @@
 import type { StoryObj, Meta } from '@storybook/react';
 
-import Header, { BackHeader, RoomSettingHeader, RoundResultHeader, TitleHeader } from './Header';
+import { RoomSettingHeader, RoundResultHeader, TitleHeader } from './components';
+import Header from './Header';
 
 const meta = {
   title: 'Header',
@@ -21,8 +22,4 @@ export const 방_설정_헤더: Story = {
 
 export const 라운드_헤더: Story = {
   render: () => <RoundResultHeader />,
-};
-
-export const 투표_현황_헤더: Story = {
-  render: () => <BackHeader title="투표 현황" />,
 };
