+++ frontend/src/components/layout/Header/components/RoundResultHeader/RoundResultHeader.tsx
@@ -1,0 +1,31 @@
+/* eslint-disable jsx-a11y/no-noninteractive-tabindex */
+import { useParams } from 'react-router-dom';
+
+import { gameTitle, headerLayout, roundText } from '../../Header.styled';
+
+import A11yOnly from '@/components/common/a11yOnly/A11yOnly';
+
+import { useBalanceContentQuery, useFocus } from '@/hooks';
+
+// 좌측 상단 라운드, 가운데 제목 차지하는 헤더 (API 호출 O) : 라운드 통계 화면
+const RoundResultHeader = () => {
+  const { roomId } = useParams();
+  const { balanceContent } = useBalanceContentQuery(Number(roomId));
+  const screenReaderRoundResult = `${balanceContent.totalRound}라운드 중. ${balanceContent.currentRound}라운드. 투표 결과 페이지`;
+  const focusRef = useFocus<HTMLElement>();
+
+  return (
+    <header css={headerLayout()} tabIndex={0} ref={focusRef}>
+      <A11yOnly>{screenReaderRoundResult}</A11yOnly>
+      <span css={roundText} aria-hidden>
+        {balanceContent.currentRound}/{balanceContent.totalRound}
+      </span>
+      <h1 css={gameTitle} aria-hidden>
+        투표 결과
+      </h1>
+      <span css={roundText} aria-hidden></span>
+    </header>
+  );
+};
+
+export default RoundResultHeader;
