+++ frontend/src/components/layout/Header/components/index.ts
@@ -1,0 +1,7 @@
+import GameHeader from './GameHeader/GameHeader';
+import MatchingResultHeader from './MatchingResultHeader/MatchingResultHeader';
+import RoomSettingHeader from './RoomSettingHeader/RoomSettingHeader';
+import RoundResultHeader from './RoundResultHeader/RoundResultHeader';
+import TitleHeader from './TitleHeader/TitleHeader';
+
+export { TitleHeader, RoomSettingHeader, RoundResultHeader, GameHeader, MatchingResultHeader };
