+++ frontend/src/pages/GamePage/GamePage.tsx
@@ -1,7 +1,7 @@
 import SelectContainer from './components/SelectContainer/SelectContainer';
 
 import Content from '@/components/layout/Content/Content';
-import { GameHeader } from '@/components/layout/Header/Header';
+import { GameHeader } from '@/components/layout/Header/components';
 import TopicContainer from '@/components/TopicContainer/TopicContainer';
 
 const GamePage = () => {
